+++ lib/apiScenario/newmanReportParser.ts
@@ -6,7 +6,13 @@ import {
   Response,
   DescriptionDefinition,
 } from "postman-collection";
-import { NewmanExecution, NewmanReport, NewmanRequest, NewmanResponse } from "./apiScenarioTypes";
+import {
+  NewmanAssertion,
+  NewmanExecution,
+  NewmanReport,
+  NewmanRequest,
+  NewmanResponse,
+} from "./apiScenarioTypes";
 
 export interface RawNewmanReport {
   run: Run;
@@ -19,10 +25,16 @@ interface Run {
   timings: { started: number; completed: number; responseAverage: number };
 }
 
+interface Assertion {
+  error?: NewmanAssertion;
+}
+
 interface RawNewmanExecution {
+  id: string;
   item: ItemDefinition;
   request: RequestDefinition;
   response: ResponseDefinition;
+  assertions?: Assertion[];
 }
 
 export function parseNewmanReport(newmanReport: RawNewmanReport): NewmanReport {
@@ -42,9 +54,11 @@ function generateExampleItem(it: RawNewmanExecution): NewmanExecution {
   const rawResp = parseResponse(resp);
   const annotation = JSON.parse((it.item.description as DescriptionDefinition)?.content || "{}");
   return {
+    id: it.id,
     request: rawReq,
     response: rawResp,
     annotation: annotation,
+    assertions: it.assertions?.map((it) => it.error!).filter((it) => it !== undefined) || [],
   };
 }