+++ lib/apiScenario/newmanReportValidator.ts
@@ -96,7 +96,7 @@ export interface NewmanReportValidatorOption extends ApiScenarioLoaderOption {
   html?: boolean;
   baseUrl?: string;
   runId?: string;
-  validationLevel?: ValidationLevel;
+  skipValidation?: boolean;
   savePayload?: boolean;
   generateExample?: boolean;
   verbose?: boolean;
@@ -117,7 +117,7 @@ export class NewmanReportValidator {
     private junitReporter: JUnitReporter
   ) {
     setDefaultOpts(this.opts, {
-      validationLevel: "validate-request-response",
+      skipValidation: false,
       savePayload: false,
       generateExample: false,
       verbose: false,
@@ -161,7 +161,9 @@ export class NewmanReportValidator {
       fileRoot: "/",
       swaggerPaths: [...this.opts.swaggerFilePaths!],
     });
-    await this.liveValidator.initialize();
+    if (!this.opts.skipValidation) {
+      await this.liveValidator.initialize();
+    }
   }
 
   public async generateReport(rawReport: NewmanReport) {
@@ -176,11 +178,16 @@ export class NewmanReportValidator {
     this.testResult.startTime = new Date(newmanReport.timings.started).toISOString();
     this.testResult.endTime = new Date(newmanReport.timings.completed).toISOString();
     this.testResult.subscriptionId = variables.subscriptionId.value as string;
+    const visitedIds = new Set<string>();
     for (const it of newmanReport.executions) {
       if (it.annotation === undefined) {
         continue;
       }
       if (it.annotation.type === "simple" || it.annotation.type === "LRO") {
+        if (visitedIds.has(it.id)) {
+          continue;
+        }
+        visitedIds.add(it.id);
         const runtimeError: RuntimeError[] = [];
         const matchedStep = this.getMatchedStep(it.annotation.step) as StepRestCall;
 
@@ -196,6 +203,15 @@ export class NewmanReportValidator {
         if (matchedStep.externalReference) {
           continue;
         }
+
+        it.assertions.forEach((assertion) => {
+          runtimeError.push({
+            code: "ASSERTION_ERROR",
+            message: `${assertion.message}`,
+            severity: "Error",
+            detail: this.dataMasker.jsonStringify(assertion.stack),
+          });
+        });
 
         const payload = this.convertToLiveValidationPayload(it);
 
@@ -225,18 +241,17 @@ export class NewmanReportValidator {
           }
           if (it.annotation.exampleName) {
             // validate real payload.
-            responseDiffResult =
-              this.opts.validationLevel === "validate-request-response"
-                ? await this.exampleResponseDiff(
-                    {
-                      step: matchedStep.step,
-                      operationId: matchedStep.operationId,
-                      example: generatedExample,
-                    },
-                    matchedStep,
-                    newmanReport
-                  )
-                : [];
+            responseDiffResult = !this.opts.skipValidation
+              ? await this.exampleResponseDiff(
+                  {
+                    step: matchedStep.step,
+                    operationId: matchedStep.operationId,
+                    example: generatedExample,
+                  },
+                  matchedStep,
+                  newmanReport
+                )
+              : [];
           }
         }
 
@@ -250,7 +265,6 @@ export class NewmanReportValidator {
             JSON.stringify(payload, null, 2)
           );
         }
-        const liveValidationResult = await this.liveValidator.validateLiveRequestResponse(payload);
 
         this.testResult.stepResult.push({
           specFilePath: matchedStep.operation._path._spec._filePath,
@@ -267,7 +281,9 @@ export class NewmanReportValidator {
           correlationId: correlationId,
           statusCode: it.response.statusCode,
           stepName: it.annotation.step,
-          liveValidationResult: liveValidationResult,
+          liveValidationResult: !this.opts.skipValidation
+            ? await this.liveValidator.validateLiveRequestResponse(payload)
+            : undefined,
         });
       }
     }
@@ -280,17 +296,27 @@ export class NewmanReportValidator {
       url: request.url.toString(),
       method: request.method.toLowerCase(),
       headers: request.headers,
-      body: request.body ? JSON.parse(request.body) : undefined,
+      body: this.parseBody(request.body),
     };
     const liveResponse: LiveResponse = {
       statusCode: response.statusCode.toString(),
       headers: response.headers,
-      body: response.body ? JSON.parse(response.body) : undefined,
+      body: this.parseBody(response.body),
     };
     return {
       liveRequest,
       liveResponse,
     };
+  }
+
+  // body may not be json string
+  private parseBody(body: string): any {
+    try {
+      return JSON.parse(body);
+    } catch (e) {
+      console.error(`Failed to parse body: ${body}`);
+      return body ? body : undefined;
+    }
   }
 
   private convertPostmanFormat<T>(obj: T, convertString: (s: string) => string): T {
@@ -440,7 +466,7 @@ export class NewmanReportValidator {
   private getRuntimeError(it: NewmanExecution): RuntimeError {
     const responseObj = this.dataMasker.jsonParse(it.response.body);
     return {
-      code: "RUNTIME_ERROR",
+      code: `${it.response.statusCode >= 500 ? "SERVER_ERROR" : "CLIENT_ERROR"}`,
       message: `statusCode: ${it.response.statusCode}, errorCode: ${responseObj?.error?.code}, errorMessage: ${responseObj?.error?.message}`,
       severity: "Error",
       detail: this.dataMasker.jsonStringify(it.response.body),