+++ lib/apiScenario/postmanCollectionGenerator.ts
@@ -23,7 +23,6 @@ import {
   ApiScenarioTestResult,
   NewmanReportValidator,
   NewmanReportValidatorOption,
-  ValidationLevel,
 } from "./newmanReportValidator";
 import { SwaggerAnalyzer, SwaggerAnalyzerOption } from "./swaggerAnalyzer";
 import { EnvironmentVariables, VariableEnv } from "./variableEnv";
@@ -53,7 +52,7 @@ export interface PostmanCollectionGeneratorOption
   generateCollection: boolean;
   baseUrl: string;
   testProxy?: string;
-  validationLevel?: ValidationLevel;
+  skipValidation?: boolean;
   savePayload?: boolean;
   generateExample?: boolean;
   skipCleanUp?: boolean;
@@ -201,11 +200,20 @@ export class PostmanCollectionGenerator {
         for (const step of report.stepResult) {
           const trafficValidationIssue: TrafficValidationIssue = {
             errors: [
-              ...step.liveValidationResult!.requestValidationResult.errors,
-              ...step.liveValidationResult!.responseValidationResult.errors,
+              ...(step.liveValidationResult?.requestValidationResult.errors ?? []),
+              ...(step.liveValidationResult?.responseValidationResult.errors ?? []),
             ],
             specFilePath: step.specFilePath,
-            operationInfo: step.liveValidationResult!.requestValidationResult.operationInfo,
+            operationInfo: step.liveValidationResult?.requestValidationResult.operationInfo ?? {
+              operationId: "unknown",
+              apiVersion: "unknown",
+            },
+          };
+
+          // mock
+          trafficValidationIssue.operationInfo!.position = {
+            line: 0,
+            column: 0,
           };
 
           if (this.opt.savePayload) {
@@ -221,7 +229,7 @@ export class PostmanCollectionGenerator {
             trafficValidationIssue.errors?.push(this.convertRuntimeException(runtimeError));
           }
 
-          if (step.liveValidationResult!.requestValidationResult.runtimeException) {
+          if (step.liveValidationResult?.requestValidationResult.runtimeException) {
             trafficValidationIssue.errors?.push(
               this.convertRuntimeException(
                 step.liveValidationResult!.requestValidationResult.runtimeException
@@ -229,7 +237,7 @@ export class PostmanCollectionGenerator {
             );
           }
 
-          if (step.liveValidationResult!.responseValidationResult.runtimeException) {
+          if (step.liveValidationResult?.responseValidationResult.runtimeException) {
             trafficValidationIssue.errors?.push(
               this.convertRuntimeException(
                 step.liveValidationResult!.responseValidationResult.runtimeException
@@ -290,6 +298,7 @@ export class PostmanCollectionGenerator {
       reportPath: path.resolve(reportExportPath, "report.html"),
       overrideLinkInReport: false,
       sdkPackage: providerNamespace,
+      markdownPath: this.opt.markdown ? path.resolve(reportExportPath, "report.md") : undefined,
     };
 
     const generator = new HtmlReportGenerator(
@@ -446,7 +455,7 @@ export class PostmanCollectionGenerator {
       html: this.opt.html,
       baseUrl: this.opt.baseUrl,
       runId: this.opt.runId,
-      validationLevel: this.opt.validationLevel,
+      skipValidation: this.opt.skipValidation,
       generateExample: this.opt.generateExample,
       savePayload: this.opt.savePayload,
       verbose: this.opt.verbose,