+++ lib/report/generateReport.ts
@@ -69,6 +69,8 @@ export class CoverageView {
   public language: string;
   public apiVersion: string = "unknown";
   public generatedDate: Date;
+  public markdownPath: string;
+  public markdown: string;
 
   public undefinedOperationCount: number = 0;
   public operationValidated: number = 0;
@@ -95,12 +97,14 @@ export class CoverageView {
     undefinedOperationCount: number = 0,
     packageName: string = "",
     language: string = "",
+    markdownPath: string = "",
     overrideLinkInReport: boolean = false,
     outputExceptionInReport: boolean = false,
     specLinkPrefix: string = "",
     payloadLinkPrefix: string = ""
   ) {
     this.package = packageName;
+    this.markdownPath = markdownPath;
     this.validationResults = validationResults;
     this.coverageResults = coverageResults;
     this.undefinedOperationCount = undefinedOperationCount;
@@ -131,6 +135,7 @@ export class CoverageView {
 
   public async prepareDataForRendering() {
     try {
+      this.markdown = await this.readMarkdown();
       const errorDefinitions = await this.loadErrorDefinitions();
       let errorsForRendering: LiveValidationIssueForRendering[];
       this.sortedValidationResults.forEach((element) => {
@@ -208,7 +213,9 @@ export class CoverageView {
       });
 
       this.resultsForRendering = this.coverageResultsForRendering.map((item) => {
-        const data = this.validationResultsForRendering.find((i) => item.spec === i.specFilePath);
+        const data = this.validationResultsForRendering.find(
+          (i) => i.specFilePath && item.spec.includes(i.specFilePath)
+        );
         return {
           ...item,
           ...data,
@@ -268,6 +275,17 @@ export class CoverageView {
       }
     } catch (e) {
       console.error(`Failed in prepareDataForRendering with err:${e?.stack};message:${e?.message}`);
+    }
+  }
+
+  private async readMarkdown() {
+    try {
+      const loader = new FileLoader({});
+      const res = await loader.load(this.markdownPath);
+      return res;
+    } catch (e) {
+      console.error(`Failed in read report.md file`);
+      return "";
     }
   }
 
@@ -389,6 +407,7 @@ export class ReportGenerator {
   private outputExceptionInReport: boolean;
   private specLinkPrefix: string;
   private payloadLinkPrefix: string;
+  private markdownPath: string;
 
   public constructor(
     validationResults: TrafficValidationIssue[],
@@ -402,6 +421,7 @@ export class ReportGenerator {
     this.reportPath = path.resolve(process.cwd(), options.reportPath!);
     this.sdkLanguage = options.sdkLanguage!;
     this.sdkPackage = options.sdkPackage!;
+    this.markdownPath = options.markdownPath!;
     this.overrideLinkInReport = options.overrideLinkInReport!;
     this.outputExceptionInReport = options.outputExceptionInReport!;
     this.specLinkPrefix = options.specLinkPrefix!;
@@ -417,6 +437,7 @@ export class ReportGenerator {
       this.undefinedOperationsCount,
       this.sdkPackage,
       this.sdkLanguage,
+      this.markdownPath,
       this.overrideLinkInReport,
       this.outputExceptionInReport,
       this.specLinkPrefix,